update chat input design and refactor components
