+++ package.json
@@ -62,6 +62,7 @@
     "axios": "^1.7.7",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.1.1",
+    "filepond": "^4.31.4",
     "filepond-plugin-file-rename": "^1.1.8",
     "filepond-plugin-file-validate-size": "^2.2.8",
     "filepond-plugin-image-exif-orientation": "^1.0.11",
